tech(api): add dryRun mode

+++ api/scripts/prod/get-answers-from-assessments.js
@@ -7,10 +7,19 @@ export class GetAnswersFromAssessments extends Script {
     super({
       description: 'Get answers from a list of assessments',
       permanent: true,
+      options: {
+        dryRun: {
+          type: 'boolean',
+          describe: 'Run the script without making any database changes',
+          default: true,
+        },
+      },
     });
   }
 
-  async handle() {
+  async handle({ logger, options }) {
+    const { dryRun } = options;
+
     const todayDate = new Date();
     const oneYearAgoString = new Date(todayDate.getFullYear() - 1, todayDate.getMonth(), todayDate.getDate())
       .toISOString()
@@ -20,15 +29,18 @@ export class GetAnswersFromAssessments extends Script {
 
     const answerToBeDeleted = await knex
       .select('answers.*')
-      .from('answers')
-      .join('assessments', 'answers.assessmentId', 'assessments.id')
+      .from('assessments')
+      .join('answers', 'answers.assessmentId', 'assessments.id')
       .whereIn('assessments.type', targetTypes)
       .where('assessments.state', 'completed')
       .whereRaw('DATE(assessments."updatedAt") = ?', [oneYearAgoString]);
 
-    const answerToBeDeletedIds = answerToBeDeleted.map(({ id }) => id);
-
-    await knex.delete().from('answers').whereIn('id', answerToBeDeletedIds);
+    if (dryRun) {
+      logger.info(`${answerToBeDeleted.length} would be inserted to parquet file`);
+    } else {
+      const answerToBeDeletedIds = answerToBeDeleted.map(({ id }) => id);
+      await knex.delete().from('answers').whereIn('id', answerToBeDeletedIds);
+    }
   }
 }
 await ScriptRunner.execute(import.meta.url, GetAnswersFromAssessments);
